Adds new pipeline of aggregation
